+++ pkg/security/probe/discarders_linux.go
@@ -9,6 +9,7 @@ import (
 	"encoding/binary"
 	"errors"
 	"fmt"
+	"github.com/DataDog/datadog-agent/pkg/security/probe/managerhelper"
 	"math"
 	"path"
 	"strings"
@@ -560,6 +561,60 @@ func dumpDiscarders(resolver *dentry.Resolver, inodeMap, statsFB, statsBB *ebpf.
 	dump.Stats = stats
 
 	return dump, nil
+}
+
+func applyDNSDefaultDropMaskFromRules(p *EBPFProbe, rs *rules.RuleSet) error {
+	bucket := rs.GetRuleBucket(model.DNSEventType.String())
+	if bucket == nil {
+		return p.setDNSDiscarderMask(^uint16(0))
+	}
+
+	var allowMask uint16
+
+	for code := 0; code < 16; code++ {
+		ev := rs.NewFakeEvent()
+		_ = ev.SetFieldValue("dns.response.code", code)
+
+		ctx := eval.NewContext(ev)
+
+		isDiscarder := true
+		for _, r := range bucket.GetRules() {
+			ok, err := r.PartialEval(ctx, "dns.response.code")
+			if err != nil {
+				var nf *eval.ErrFieldNotFound
+				if errors.As(err, &nf) {
+					continue
+				}
+				isDiscarder = false
+				break
+			}
+			if ok {
+				isDiscarder = false
+				break
+			}
+		}
+
+		if !isDiscarder {
+			allowMask |= 1 << uint(code)
+		}
+	}
+
+	return p.setDNSDiscarderMask(^allowMask)
+}
+
+func (p *EBPFProbe) setDNSDiscarderMask(dnsMask uint16) error {
+	bufferSelector, err := managerhelper.Map(p.Manager, "filtered_dns_rcodes")
+	if err != nil {
+		return err
+	}
+
+	err = bufferSelector.Put(uint32(0), dnsMask)
+	if err != nil {
+		return err
+	}
+
+	seclog.Tracef("DNS discarder for response code: %d", dnsMask)
+	return nil
 }
 
 func init() {
